added new methods from the recent w3c java binding

+++ java/dom/jni/MouseEventImpl.java
@@ -53,111 +53,111 @@ public class MouseEventImpl extends UIEventImpl implements MouseEvent {
             " p=" + Long.toHexString(p_nsIDOMEvent) + ">";
     }
 
-  /**
-   * <code>screenX</code> indicates the horizontal coordinate at which the 
-   * event occurred in relative to the origin of the screen coordinate system.
-   */
-  public native int                getScreenX();
-
-  /**
-   * <code>screenY</code> indicates the vertical coordinate at which the event 
-   * occurred relative to the origin of the screen coordinate system.
-   */
-  public native int                getScreenY();
-
-  /**
-   * <code>clientX</code> indicates the horizontal coordinate at which the 
-   * event occurred relative to the DOM implementation's client area.
-   */
-  public native int                getClientX();
-
-  /**
-   * <code>clientY</code> indicates the vertical coordinate at which the event 
-   * occurred relative to the DOM implementation's client area.
-   */
-  public native int                getClientY();
-
-  /**
-   * <code>ctrlKey</code> indicates whether the 'ctrl' key was depressed 
-   * during the firing of the event.
-   */
-  public native boolean            getCtrlKey();
-
-  /**
-   * <code>shiftKey</code> indicates whether the 'shift' key was depressed 
-   * during the firing of the event.
-   */
-  public native boolean            getShiftKey();
-
-  /**
-   * <code>altKey</code> indicates whether the 'alt' key was depressed during 
-   * the firing of the event.  On some platforms this key may map to an 
-   * alternative key name.
-   */
-  public native boolean            getAltKey();
-
-  /**
-   * <code>metaKey</code> indicates whether the 'meta' key was depressed 
-   * during the firing of the event.  On some platforms this key may map to 
-   * an alternative key name.
-   */
-  public native boolean            getMetaKey();
-
-  /**
-   * During mouse events caused by the depression or release of a mouse 
-   * button, <code>button</code> is used to indicate which mouse button 
-   * changed state.
-   */
-  public native short              getButton();
-
-  /**
-   * <code>relatedNode</code> is used to identify a secondary node related to 
-   * a UI event.
-   */
-  public Node               getRelatedNode() {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
-   * 
-   * @param typeArg Specifies the event type.
-   * @param canBubbleArg Specifies whether or not the event can bubble.
-   * @param cancelableArg Specifies whether or not the event's default  action 
-   *   can be prevent.
-   * @param viewArg Specifies the <code>Event</code>'s 
-   *   <code>AbstractView</code>.
-   * @param detailArg Specifies the <code>Event</code>'s mouse click count.
-   * @param screenXArg Specifies the <code>Event</code>'s screen x coordinate
-   * @param screenYArg Specifies the <code>Event</code>'s screen y coordinate
-   * @param clientXArg Specifies the <code>Event</code>'s client x coordinate
-   * @param clientYArg Specifies the <code>Event</code>'s client y coordinate
-   * @param ctrlKeyArg Specifies whether or not control key was depressed 
-   *   during the <code>Event</code>.
-   * @param altKeyArg Specifies whether or not alt key was depressed during 
-   *   the  <code>Event</code>.
-   * @param shiftKeyArg Specifies whether or not shift key was depressed 
-   *   during the <code>Event</code>.
-   * @param metaKeyArg Specifies whether or not meta key was depressed during 
-   *   the  <code>Event</code>.
-   * @param buttonArg Specifies the <code>Event</code>'s mouse button.
-   * @param relatedNodeArg Specifies the <code>Event</code>'s related Node.
-   */
-  public void               initMouseEvent(String typeArg, 
-                                           boolean canBubbleArg, 
-                                           boolean cancelableArg, 
-                                           AbstractView viewArg, 
-                                           short detailArg, 
-                                           int screenXArg, 
-                                           int screenYArg, 
-                                           int clientXArg, 
-                                           int clientYArg, 
-                                           boolean ctrlKeyArg, 
-                                           boolean altKeyArg, 
-                                           boolean shiftKeyArg, 
-                                           boolean metaKeyArg, 
-                                           short buttonArg, 
-                                           Node relatedNodeArg) {
-    throw new UnsupportedOperationException();
-  }
+    /**
+     * <code>screenX</code> indicates the horizontal coordinate at which the 
+     * event occurred in relative to the origin of the screen coordinate system.
+     */
+    public native int                getScreenX();
+    
+    /**
+     * <code>screenY</code> indicates the vertical coordinate at which the event 
+     * occurred relative to the origin of the screen coordinate system.
+     */
+    public native int                getScreenY();
+    
+    /**
+     * <code>clientX</code> indicates the horizontal coordinate at which the 
+     * event occurred relative to the DOM implementation's client area.
+     */
+    public native int                getClientX();
+    
+    /**
+     * <code>clientY</code> indicates the vertical coordinate at which the event 
+     * occurred relative to the DOM implementation's client area.
+     */
+    public native int                getClientY();
+    
+    /**
+     * <code>ctrlKey</code> indicates whether the 'ctrl' key was depressed 
+     * during the firing of the event.
+     */
+    public native boolean            getCtrlKey();
+    
+    /**
+     * <code>shiftKey</code> indicates whether the 'shift' key was depressed 
+     * during the firing of the event.
+     */
+    public native boolean            getShiftKey();
+    
+    /**
+     * <code>altKey</code> indicates whether the 'alt' key was depressed during 
+     * the firing of the event.  On some platforms this key may map to an 
+     * alternative key name.
+     */
+    public native boolean            getAltKey();
+    
+    /**
+     * <code>metaKey</code> indicates whether the 'meta' key was depressed 
+     * during the firing of the event.  On some platforms this key may map to 
+     * an alternative key name.
+     */
+    public native boolean            getMetaKey();
+    
+    /**
+     * During mouse events caused by the depression or release of a mouse 
+     * button, <code>button</code> is used to indicate which mouse button 
+     * changed state.
+     */
+    public native short              getButton();
+    
+    /**
+     * <code>relatedNode</code> is used to identify a secondary node related to 
+     * a UI event.
+     */
+    public Node               getRelatedNode() {
+	throw new UnsupportedOperationException();
+    }
+    
+    /**
+     * 
+     * @param typeArg Specifies the event type.
+     * @param canBubbleArg Specifies whether or not the event can bubble.
+     * @param cancelableArg Specifies whether or not the event's default  action 
+     *   can be prevent.
+     * @param viewArg Specifies the <code>Event</code>'s 
+     *   <code>AbstractView</code>.
+     * @param detailArg Specifies the <code>Event</code>'s mouse click count.
+     * @param screenXArg Specifies the <code>Event</code>'s screen x coordinate
+     * @param screenYArg Specifies the <code>Event</code>'s screen y coordinate
+     * @param clientXArg Specifies the <code>Event</code>'s client x coordinate
+     * @param clientYArg Specifies the <code>Event</code>'s client y coordinate
+     * @param ctrlKeyArg Specifies whether or not control key was depressed 
+     *   during the <code>Event</code>.
+     * @param altKeyArg Specifies whether or not alt key was depressed during 
+     *   the  <code>Event</code>.
+     * @param shiftKeyArg Specifies whether or not shift key was depressed 
+     *   during the <code>Event</code>.
+     * @param metaKeyArg Specifies whether or not meta key was depressed during 
+     *   the  <code>Event</code>.
+     * @param buttonArg Specifies the <code>Event</code>'s mouse button.
+     * @param relatedNodeArg Specifies the <code>Event</code>'s related Node.
+     */
+    public void initMouseEvent(String typeArg, 
+                               boolean canBubbleArg, 
+                               boolean cancelableArg, 
+                               AbstractView viewArg, 
+                               int detailArg, 
+                               int screenXArg, 
+                               int screenYArg, 
+                               int clientXArg, 
+                               int clientYArg, 
+                               boolean ctrlKeyArg, 
+                               boolean altKeyArg, 
+                               boolean shiftKeyArg, 
+                               boolean metaKeyArg, 
+                               short buttonArg, 
+                               Node relatedNodeArg) {
+        throw new UnsupportedOperationException();
+    }
 }
